pyproject.toml
# see https://flit.pypa.io/en/latest/pyproject_toml.html
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scan2wiki"
description = "scan2wiki - nicegui based DMS with option to scan documents to a semantic mediawiki"
home-page = "https://github.com/WolfgangFahl/scan2wiki"
keywords = [ "nicegui", "scan", "DMS", "OCR","mediawiki" ]

authors = [
  {name = "Wolfgang Fahl", email = "wf@WolfgangFahl.com"}
]

maintainers = [
  { name = "Wolfgang Fahl", email = "wf@WolfgangFahl.com" },
]

readme = "README.md"
license = {text = "Apache-2.0"}
dependencies = [
	# https://pypi.org/project/pybasemkit/
	# https://github.com/WolfgangFahl/pybasemkit
	# Python base module kit: YAML/JSON I/O, structured logging, CLI tooling, shell execution, and remote pydevd debug support.
	"pybasemkit>=0.0.3",
	# https://github.com/WolfgangFahl/nicegui_widgets
	"ngwidgets>=0.27.1",
	# https://pypi.org/project/dataclasses-json/
	"dataclasses-json>=0.6.1",
	# Mediawiki and SMW API
	# https://pypi.org/project/py-3rdparty-mediawiki/
    'py-3rdparty-mediawiki>=0.15.4',
	# https://pypi.org/project/pylodstorage/
	"pylodstorage>=0.16.4",
	# https://github.com/claird/PyPDF4
	# https://pypi.org/project/PyPDF2/
	# issue https://github.com/sfneal/PyPDF3/issues/13
	# https://github.com/sfneal/PyPDF3
	# makes it unusuable (showstopper) for our usecase
	# PyPDF3
	# https://pypi.org/project/pdfminer/
	# https://github.com/pdfminer/pdfminer.six
	# "pdfminer.six",
	# https://github.com/pymupdf/pymupdf
	# https://pypi.org/project/PyMuPDF/
	"PyMuPDF>=1.25.5",
	# https://pypi.org/project/watchdog/
	"watchdog",
	# https://pypi.org/project/APScheduler/
	"APScheduler",
	# https://pypi.org/project/beautifulsoup4/
	"beautifulsoup4>=4.11.1",
	# https://github.com/NaturalHistoryMuseum/pyzbar
	# https://pypi.org/project/pyzbar/
	"pyzbar>=0.1.9"
]

requires-python = ">=3.10"
classifiers=[
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: User Interfaces",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License"
]
dynamic = ["version"]
[tool.hatch.version]
path = "scan/__init__.py"

[project.urls]
Home = "https://github.com/WolfgangFahl/scan2wiki"
Documentation = "https://wiki.bitplan.com/index.php/scan2wiki"
Source = "https://github.com/WolfgangFahl/scan2wiki"

[project.optional-dependencies]
test = [
  "green",
]

[tool.hatch.build.targets.wheel]
only-include = ["scan","scan2wiki_examples"]

[tool.hatch.build.targets.wheel.sources]
"scan" = "scan"
"scan2wiki_examples" = "scan2wiki_examples"

[project.scripts]
scan2wiki = "scan.scan_cmd:main"
